refactor: move PageLoader to animations directory
