+++ packages/client/src/elements/animations/PageLoader/PageLoader.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
 
-import PageLoader from '.';
+import PageLoader from './index';
 
 export default {
-  title: 'Elements/Page Loader',
+  title: 'Elements/animations/Page Loader',
   component: PageLoader,
 } as Meta;
