+++ packages/client/src/elements/animations/index.ts
@@ -1,2 +1,3 @@
 export { default as AnimatedLogo } from './AnimatedLogo';
+export { default as PageLoader } from './PageLoader';
 export { default as Spinner } from './Spinner';
