+++ packages/client/src/elements/index.ts
@@ -1,4 +1,4 @@
-export { AnimatedLogo, Spinner } from './animations';
+export { AnimatedLogo, PageLoader, Spinner } from './animations';
 export { default as BackgroundSideLogo } from './BackgroundSideLogo';
 export { Button, IconButton } from './buttons';
 export { default as Card } from './Card';
@@ -7,5 +7,4 @@ export { default as Form } from './Form';
 export { Checkbox, TextField } from './inputs';
 export { DefaultLayout, LayoutContainer } from './layouts';
 export { default as Link } from './Link';
-export { default as PageLoader } from './PageLoader';
 export { default as Snackbar } from './Snackbar';
